* Log actions in reactotron * Fix for store - exchanger was creating a second store. Refactored in order to use one instance of createStore()
